+++ tests/TestCase/Mailer/Transport/SmtpTransportTest.php
@@ -16,50 +16,9 @@ declare(strict_types=1);
 namespace Cake\Test\TestCase\Mailer\Transport;
 
 use Cake\Mailer\Email;
-use Cake\Mailer\Transport\SmtpTransport;
 use Cake\Network\Exception\SocketException;
-use Cake\Network\Socket;
 use Cake\TestSuite\TestCase;
-
-/**
- * Help to test SmtpTransport
- */
-class SmtpTestTransport extends SmtpTransport
-{
-    /**
-     * Helper to change the socket
-     *
-     * @param Socket $socket
-     * @return void
-     */
-    public function setSocket(Socket $socket)
-    {
-        $this->_socket = $socket;
-    }
-
-    /**
-     * Disabled the socket change
-     *
-     * @return void
-     */
-    protected function _generateSocket(): void
-    {
-    }
-
-    /**
-     * Magic function to call protected methods
-     *
-     * @param string $method
-     * @param array $args
-     * @return mixed
-     */
-    public function __call($method, $args)
-    {
-        $method = '_' . $method;
-
-        return call_user_func_array([$this, $method], $args);
-    }
-}
+use TestApp\Mailer\Transport\SmtpTestTransport;
 
 /**
  * Test case
@@ -67,7 +26,7 @@ class SmtpTestTransport extends SmtpTransport
 class SmtpTransportTest extends TestCase
 {
     /**
-     * @var \Cake\Test\TestCase\Mailer\Transport\SmtpTestTransport
+     * @var \TestApp\Mailer\Transport\SmtpTestTransport
      */
     protected $SmtpTransport;
