+++ tests/TestCase/ORM/BehaviorTest.php
@@ -15,165 +15,12 @@ declare(strict_types=1);
  */
 namespace Cake\Test\TestCase\ORM;
 
-use Cake\ORM\Behavior;
+use Cake\Core\Exception\Exception;
+use Cake\ORM\Table;
 use Cake\TestSuite\TestCase;
-
-/**
- * Test Stub.
- */
-class TestBehavior extends Behavior
-{
-    /**
-     * Test for event bindings.
-     */
-    public function beforeFind()
-    {
-    }
-
-    /**
-     * Test for event bindings.
-     */
-    public function beforeRules()
-    {
-    }
-
-    /**
-     * Test for event bindings.
-     */
-    public function afterRules()
-    {
-    }
-
-    /**
-     * Test for event bindings.
-     */
-    public function buildRules()
-    {
-    }
-
-    /**
-     * Test for event bindings.
-     */
-    public function afterSaveCommit()
-    {
-    }
-
-    /**
-     * Test for event bindings.
-     */
-    public function afterDeleteCommit()
-    {
-    }
-}
-
-/**
- * Test Stub.
- */
-class Test2Behavior extends Behavior
-{
-    protected $_defaultConfig = [
-        'implementedFinders' => [
-            'foo' => 'findFoo',
-        ],
-        'implementedMethods' => [
-            'doSomething' => 'doSomething',
-        ],
-    ];
-
-    /**
-     * Test for event bindings.
-     */
-    public function beforeFind()
-    {
-    }
-
-    /**
-     * Test finder
-     */
-    public function findFoo()
-    {
-    }
-
-    /**
-     * Test method
-     */
-    public function doSomething()
-    {
-    }
-}
-
-/**
- * Test3Behavior
- */
-class Test3Behavior extends Behavior
-{
-    /**
-     * Test for event bindings.
-     */
-    public function beforeFind()
-    {
-    }
-
-    /**
-     * Test finder
-     */
-    public function findFoo()
-    {
-    }
-
-    /**
-     * Test method
-     */
-    public function doSomething()
-    {
-    }
-
-    /**
-     * Test method to ensure it is ignored as a callable method.
-     */
-    public function verifyConfig(): void
-    {
-        parent::verifyConfig();
-    }
-
-    /**
-     * implementedEvents
-     *
-     * This class does pretend to implement beforeFind
-     *
-     * @return array
-     */
-    public function implementedEvents(): array
-    {
-        return ['Model.beforeFind' => 'beforeFind'];
-    }
-
-    /**
-     * implementedFinders
-     */
-    public function implementedFinders(): array
-    {
-    }
-
-    /**
-     * implementedMethods
-     */
-    public function implementedMethods(): array
-    {
-    }
-
-    /**
-     * Expose protected method for testing
-     *
-     * Since this is public - it'll show up as callable which is a side-effect
-     *
-     * @return array
-     */
-    public function testReflectionCache()
-    {
-        return $this->_reflectionCache();
-    }
-}
+use TestApp\Model\Behavior\Test2Behavior;
+use TestApp\Model\Behavior\Test3Behavior;
+use TestApp\Model\Behavior\TestBehavior;
 
 /**
  * Behavior test case
@@ -187,7 +34,7 @@ class BehaviorTest extends TestCase
      */
     public function testConstructor()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $config = ['key' => 'value'];
         $behavior = new TestBehavior($table, $config);
         $this->assertEquals($config, $behavior->getConfig());
@@ -200,7 +47,7 @@ class BehaviorTest extends TestCase
      */
     public function testGetTable()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
 
         $behavior = new TestBehavior($table);
         $this->assertSame($table, $behavior->getTable());
@@ -208,7 +55,7 @@ class BehaviorTest extends TestCase
 
     public function testReflectionCache()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test3Behavior($table);
         $expected = [
             'finders' => [
@@ -229,7 +76,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedEvents()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new TestBehavior($table);
         $expected = [
             'Model.beforeFind' => 'beforeFind',
@@ -249,7 +96,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedEventsWithPriority()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new TestBehavior($table, ['priority' => 10]);
         $expected = [
             'Model.beforeFind' => [
@@ -287,7 +134,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedMethods()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table);
         $expected = [
             'doSomething' => 'doSomething',
@@ -302,7 +149,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedMethodsAliased()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedMethods' => [
                 'aliased' => 'doSomething',
@@ -321,7 +168,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedMethodsDisabled()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedMethods' => [],
         ]);
@@ -336,7 +183,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedFinders()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table);
         $expected = [
             'foo' => 'findFoo',
@@ -351,7 +198,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedFindersAliased()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedFinders' => [
                 'aliased' => 'findFoo',
@@ -370,7 +217,7 @@ class BehaviorTest extends TestCase
      */
     public function testImplementedFindersDisabled()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedFinders' => [],
         ]);
@@ -386,7 +233,7 @@ class BehaviorTest extends TestCase
      */
     public function testVerifyConfig()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table);
         $behavior->verifyConfig();
         $this->assertTrue(true, 'No exception thrown');
@@ -401,7 +248,7 @@ class BehaviorTest extends TestCase
      */
     public function testVerifyConfigImplementedFindersOverridden()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedFinders' => [
                 'aliased' => 'findFoo',
@@ -421,7 +268,7 @@ class BehaviorTest extends TestCase
     {
         $this->expectException(\Cake\Core\Exception\Exception::class);
         $this->expectExceptionMessage('The method findNotDefined is not callable on class Cake\Test\TestCase\ORM\Test2Behavior');
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedFinders' => [
                 'aliased' => 'findNotDefined',
@@ -439,7 +286,7 @@ class BehaviorTest extends TestCase
      */
     public function testVerifyConfigImplementedMethodsOverridden()
     {
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table);
         $behavior = new Test2Behavior($table, [
             'implementedMethods' => [
@@ -458,9 +305,9 @@ class BehaviorTest extends TestCase
      */
     public function testVerifyImplementedMethodsInvalid()
     {
-        $this->expectException(\Cake\Core\Exception\Exception::class);
+        $this->expectException(Exception::class);
         $this->expectExceptionMessage('The method iDoNotExist is not callable on class Cake\Test\TestCase\ORM\Test2Behavior');
-        $table = $this->getMockBuilder('Cake\ORM\Table')->getMock();
+        $table = $this->getMockBuilder(Table::class)->getMock();
         $behavior = new Test2Behavior($table, [
             'implementedMethods' => [
                 'aliased' => 'iDoNotExist',
