+++ tests/TestCase/Utility/MergeVariablesTraitTest.php
@@ -18,62 +18,6 @@ namespace Cake\Test\TestCase\Utility;
 use Cake\TestSuite\TestCase;
 use Cake\Utility\MergeVariablesTrait;
 
-class Base
-{
-    use MergeVariablesTrait;
-
-    public $hasBoolean = false;
-
-    public $listProperty = ['One'];
-
-    public $assocProperty = ['Red'];
-
-    public function mergeVars($properties, $options = [])
-    {
-        return $this->_mergeVars($properties, $options);
-    }
-}
-
-class Child extends Base
-{
-    public $hasBoolean = ['test'];
-
-    public $listProperty = ['Two', 'Three'];
-
-    public $assocProperty = [
-        'Green' => ['lime'],
-        'Orange',
-    ];
-
-    public $nestedProperty = [
-        'Red' => [
-            'apple' => 'gala',
-        ],
-        'Green' => [
-            'citrus' => 'lime',
-        ],
-    ];
-}
-
-class Grandchild extends Child
-{
-    public $listProperty = ['Four', 'Five'];
-
-    public $assocProperty = [
-        'Green' => ['apple'],
-        'Yellow' => ['banana'],
-    ];
-
-    public $nestedProperty = [
-        'Red' => [
-            'citrus' => 'blood orange',
-        ],
-        'Green' => [
-            'citrus' => 'key lime',
-        ],
-    ];
-}
-
 /**
  * MergeVariablesTrait test case
  */
@@ -166,4 +110,60 @@ class MergeVariablesTraitTest extends TestCase
         $object->mergeVars(['hasBoolean']);
         $this->assertEquals(['test'], $object->hasBoolean);
     }
+}
+
+class Base
+{
+    use MergeVariablesTrait;
+
+    public $hasBoolean = false;
+
+    public $listProperty = ['One'];
+
+    public $assocProperty = ['Red'];
+
+    public function mergeVars($properties, $options = [])
+    {
+        return $this->_mergeVars($properties, $options);
+    }
+}
+
+class Child extends Base
+{
+    public $hasBoolean = ['test'];
+
+    public $listProperty = ['Two', 'Three'];
+
+    public $assocProperty = [
+        'Green' => ['lime'],
+        'Orange',
+    ];
+
+    public $nestedProperty = [
+        'Red' => [
+            'apple' => 'gala',
+        ],
+        'Green' => [
+            'citrus' => 'lime',
+        ],
+    ];
+}
+
+class Grandchild extends Child
+{
+    public $listProperty = ['Four', 'Five'];
+
+    public $assocProperty = [
+        'Green' => ['apple'],
+        'Yellow' => ['banana'],
+    ];
+
+    public $nestedProperty = [
+        'Red' => [
+            'citrus' => 'blood orange',
+        ],
+        'Green' => [
+            'citrus' => 'key lime',
+        ],
+    ];
 }
