+++ tests/test_app/TestApp/Mailer/Stub.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Mailer;
+
+use Cake\Mailer\MailerAwareTrait;
+
+class Stub
+{
+    use MailerAwareTrait {
+        getMailer as public;
+    }
+}
