+++ tests/test_app/TestApp/Model/Behavior/TestBehavior.php
@@ -1,0 +1,50 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Behavior;
+
+use Cake\ORM\Behavior;
+
+class TestBehavior extends Behavior
+{
+    /**
+     * Test for event bindings.
+     */
+    public function beforeFind()
+    {
+    }
+
+    /**
+     * Test for event bindings.
+     */
+    public function beforeRules()
+    {
+    }
+
+    /**
+     * Test for event bindings.
+     */
+    public function afterRules()
+    {
+    }
+
+    /**
+     * Test for event bindings.
+     */
+    public function buildRules()
+    {
+    }
+
+    /**
+     * Test for event bindings.
+     */
+    public function afterSaveCommit()
+    {
+    }
+
+    /**
+     * Test for event bindings.
+     */
+    public function afterDeleteCommit()
+    {
+    }
+}
