+++ tests/test_app/TestApp/Model/Entity/ProtectedArticle.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Entity;
+
+use Cake\ORM\Entity;
+
+/**
+ * Test entity for mass assignment.
+ */
+class ProtectedArticle extends Entity
+{
+    protected $_accessible = [
+        'title' => true,
+        'body' => true,
+    ];
+}
