+++ tests/test_app/TestApp/Model/Entity/ProtectedEntity.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Entity;
+
+use Cake\ORM\Entity;
+
+class ProtectedEntity extends Entity
+{
+    protected $_accessible = [
+        'id' => true,
+        'title' => false,
+    ];
+}
