+++ tests/test_app/TestApp/Model/Entity/ProtectedUser.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Entity;
+
+use Cake\ORM\Entity;
+
+/**
+ * Entity for testing with hidden fields.
+ */
+class ProtectedUser extends Entity
+{
+    protected $_hidden = ['password'];
+}
