+++ tests/test_app/TestApp/Model/Entity/TranslateBakedArticle.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Entity;
+
+use Cake\ORM\Behavior\Translate\TranslateTrait;
+use Cake\ORM\Entity;
+
+class TranslateBakedArticle extends Entity
+{
+    use TranslateTrait;
+
+    protected $_accessible = [
+        'title' => true,
+        'body' => true,
+    ];
+}
