+++ tests/test_app/TestApp/Model/Table/UsersTable.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Table;
+
+use Cake\ORM\Table;
+
+/**
+ * Used to test correct class is instantiated when using $this->getTableLocator()->get();
+ */
+class UsersTable extends Table
+{
+}
